backend/requirements.txt
fastapi
uvicorn
openai
chromadb
unstructured[all-docs]
python-dotenv
pydantic
requests
